+++ Plugins/UVOCBot.Plugins.Planetside/PlanetsidePluginOptions.cs
@@ -3,21 +3,18 @@
 public record PlanetsidePluginOptions
 {
     /// <summary>
-    /// Gets or sets the endpoint at which the fisu's API can be found.
+    /// Gets the endpoint at which the fisu API can be found.
     /// </summary>
     public string FisuApiEndpoint { get; init; }
 
-    public string HonuApiEndpoint { get; init; }
-
     /// <summary>
-    /// Gets or sets the key used to connect to Daybreak Game's Census API.
+    /// Gets the endpoint at which the honu API can be found.
     /// </summary>
-    public string CensusApiKey { get; init; }
+    public string HonuApiEndpoint { get; init; }
 
     public PlanetsidePluginOptions()
     {
         FisuApiEndpoint = "https://ps2.fisu.pw/api";
         HonuApiEndpoint = "https://wt.honu.pw/api";
-        CensusApiKey = "example";
     }
 }