+++ Plugins/UVOCBot.Plugins.Planetside/Services/CachingCensusApiService.cs
@@ -1,4 +1,5 @@
 ﻿using DbgCensus.Core.Objects;
+using DbgCensus.EventStream.Objects.Events.Worlds;
 using DbgCensus.Rest.Abstractions;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Logging;
@@ -20,10 +21,12 @@ public class CachingCensusApiService : CensusApiService
 {
     private readonly IMemoryCache _cache;
 
-    public CachingCensusApiService(
+    public CachingCensusApiService
+    (
         ILogger<CachingCensusApiService> logger,
         IQueryService queryService,
-        IMemoryCache cache)
+        IMemoryCache cache
+    )
         : base(logger, queryService)
     {
         _cache = cache;
@@ -115,6 +118,16 @@ public class CachingCensusApiService : CensusApiService
         }
 
         return maps;
+    }
 
+    public override async Task<Result<MetagameEvent>> GetMetagameEventAsync(ValidWorldDefinition world, ValidZoneDefinition zone, CancellationToken ct = default)
+    {
+        if (_cache.TryGetValue(CacheKeyHelpers.GetMetagameEventKey((WorldDefinition)world, (ZoneDefinition)zone), out MetagameEvent found))
+            return found;
+
+        // Note that we don't cache the result here
+        // This is because we expect the MetagameEventResponder
+        // to keep events up-to-date in a more reliable manner.
+        return await base.GetMetagameEventAsync(world, zone, ct);
     }
 }