+++ UVOCBot/Config/GeneralOptions.cs
@@ -10,11 +10,6 @@ public record GeneralOptions
     public string BotToken { get; init; }
 
     /// <summary>
-    /// Gets or sets the default command prefix.
-    /// </summary>
-    public string CommandPrefix { get; init; }
-
-    /// <summary>
     /// The text to show in the custom status area of the bot profile. Will always be prefixed by the 'Playing' operator.
     /// </summary>
     public string DiscordPresence { get; init; }
@@ -27,7 +22,6 @@ public record GeneralOptions
     public GeneralOptions()
     {
         BotToken = string.Empty;
-        CommandPrefix = "<>"; // No prefix, most commands use the slash infrastructure
         DebugGuildIds = new List<ulong>();
         DiscordPresence = string.Empty;
     }